♻️(AddTagModal): create component addTagModal
