+++ frontend/src/pages/Tags/components/AddTagModal/AddTagModal.wrap.tsx
@@ -1,20 +1,19 @@
 import { connect } from 'react-redux';
 import { RootState } from 'redux/types';
-import Tags from './Tags';
+import { AddTagModal } from './AddTagModal';
 import { Dispatch } from 'redux';
-import { loadTags } from 'redux/Tag/tag.actions';
-import { getTags, isTagLoading } from 'redux/Tag/tag.selectors';
+import { addTag } from 'redux/Tag/tag.actions';
+import { isTagLoading } from 'redux/Tag/tag.selectors';
 
 const mapStateToProps = (state: RootState) => ({
-  tags: getTags(state),
   isTagLoading: isTagLoading(state),
 });
 
 const mapDispatchToProps = (dispatch: Dispatch) => ({
-  loadTags: () => dispatch(loadTags.request({})),
+  addTag: (code: string, description: string) => dispatch(addTag.request({ code, description })),
 });
 
 export default connect(
   mapStateToProps,
   mapDispatchToProps,
-)(Tags);
+)(AddTagModal);
