+++ frontend/src/pages/Tags/components/AddTagModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AddTagModal.wrap';
